refactor: rename com.wanderersoftherift.wotr.codec to com.wanderersoftherift.wotr.seralization, moved WotrAttachments.AttachmentSerializerFromDataCodec to com.wanderersoftherift.wotr.seralization, (re)created AbilitySlots.SERIALIZER field
