+++ src/main/java/com/wanderersoftherift/wotr/item/crafting/KeyForgeRecipe.java
@@ -4,7 +4,6 @@ import com.mojang.datafixers.util.Pair;
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.MapCodec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
-import com.wanderersoftherift.wotr.codec.DispatchedPair;
 import com.wanderersoftherift.wotr.init.WotrBlocks;
 import com.wanderersoftherift.wotr.init.WotrItems;
 import com.wanderersoftherift.wotr.init.recipe.WotrRecipeCategories;
@@ -12,6 +11,7 @@ import com.wanderersoftherift.wotr.init.recipe.WotrRecipeSerializers;
 import com.wanderersoftherift.wotr.init.recipe.WotrRecipeTypes;
 import com.wanderersoftherift.wotr.item.crafting.display.EssenceSlotDisplay;
 import com.wanderersoftherift.wotr.item.crafting.display.KeyForgeRecipeDisplay;
+import com.wanderersoftherift.wotr.serialization.DispatchedPair;
 import net.minecraft.advancements.Advancement;
 import net.minecraft.advancements.AdvancementHolder;
 import net.minecraft.advancements.AdvancementRequirements;
