+++ src/main/java/com/wanderersoftherift/wotr/world/level/levelgen/processor/AttachmentProcessor.java
@@ -3,7 +3,7 @@ package com.wanderersoftherift.wotr.world.level.levelgen.processor;
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.MapCodec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
-import com.wanderersoftherift.wotr.codec.OutputStateCodecs;
+import com.wanderersoftherift.wotr.serialization.OutputStateCodecs;
 import com.wanderersoftherift.wotr.util.EnumEntries;
 import com.wanderersoftherift.wotr.world.level.levelgen.processor.util.ProcessorUtil;
 import com.wanderersoftherift.wotr.world.level.levelgen.processor.util.StructureRandomType;
